fix : 순환참조 수정
