+++ src/main/java/com/example/tnote/boundedContext/user/service/AuthService.java
@@ -3,9 +3,12 @@ package com.example.tnote.boundedContext.user.service;
 import com.example.tnote.base.exception.CustomException;
 import com.example.tnote.base.utils.JwtTokenProvider;
 import com.example.tnote.boundedContext.RefreshToken.entity.RefreshToken;
+import com.example.tnote.boundedContext.RefreshToken.repository.RefreshTokenRepository;
 import com.example.tnote.boundedContext.RefreshToken.service.RefreshTokenService;
 import com.example.tnote.boundedContext.user.dto.JwtResponse;
+import com.example.tnote.boundedContext.user.dto.KakaoUnlinkResponse;
 import com.example.tnote.boundedContext.user.dto.Token;
+import com.example.tnote.boundedContext.user.dto.UserDeleteResponseDto;
 import com.example.tnote.boundedContext.user.entity.User;
 import com.example.tnote.boundedContext.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +26,7 @@ public class AuthService {
     private final RefreshTokenService refreshTokenService;
     private final UserRepository userRepository;
     private final JwtTokenProvider jwtTokenProvider;
+    private final RefreshTokenRepository refreshTokenRepository;
 
     @Transactional
     public JwtResponse redirect(String provider, String code, String state) {
@@ -58,4 +62,28 @@ public class AuthService {
                 .userId(userId)
                 .build();
     }
+
+    @Transactional
+    public UserDeleteResponseDto deleteUser(Long userId, String code) {
+
+        User currentUser = userRepository.findById(userId)
+                .orElseThrow(() -> CustomException.USER_NOT_FOUND);
+
+        deleteAll(currentUser);
+
+        String accessToken = kakaoRequestService.getToken(code).getAccessToken();
+
+        KakaoUnlinkResponse unlink = kakaoRequestService.unLink(accessToken);
+
+        return UserDeleteResponseDto.builder()
+                .id(unlink.getId())
+                .build();
+    }
+
+    // 연관키로 묶여 있음
+    private void deleteAll(User currentUser) {
+        refreshTokenRepository.deleteByKeyEmail(currentUser.getEmail());
+        userRepository.delete(currentUser);
+    }
+
 }
